requirements.txt
streamlit
pandas
plotly
openpyxl
mysql-connector-python
matplotlib
